converted from float to double

+++ src/geo.cpp
@@ -10,11 +10,11 @@
 namespace py =pybind11;
 // using namespace std;
 
-void split(std::string &, float *,float*);
-float dmstodd(char *);
-float distance(float a1,float b1,float a2,float b2){
+void split(std::string &, double *,double*);
+double dmstodd(char *);
+double distance(double a1,double b1,double a2,double b2){
     // return abs(a1-a2)+abs(b1-b2);
-    float da,db,c;
+    double da,db,c;
     a1*=(PI/180);
     a2*=(PI/180);
     b1*=(PI/180);
@@ -27,28 +27,28 @@ float distance(float a1,float b1,float a2,float b2){
     return c;
 
 }
-float calculateDistance(std::string &v1, std::string &v2){
-    float a1,b1,a2,b2;
+double calculateDistance(std::string &v1, std::string &v2){
+    double a1,b1,a2,b2;
     split(v1,&a1,&b1);split(v2,&a2,&b2);
     return distance(a1,b1,a2,b2);
 }
 class query{
     public:
-    std::vector<std::array<float, 2>> mas;
+    std::vector<std::array<double, 2>> mas;
     query(const py::list & abc){
         ;
         for( auto &i:abc){
-            std::array<float,2> a;
+            std::array<double,2> a;
             std::string v =py::cast<std::string >(i);
             split(v,&a[0],&a[1]);
             this->mas.push_back(a);
         }
     }
     py::tuple nearest(std::string & a){
-        float lt,ln,t;
+        double lt,ln,t;
         int ind=0,count=0;
         split(a,&lt,&ln);
-        float min=10000;
+        double min=10000;
         // for (int i=0;i<this->mas.size();i++){
         for(auto &i:this->mas){
             t=distance(i[0],i[1],lt,ln);
@@ -60,8 +60,8 @@ class query{
         return py::make_tuple(ind,this->mas[ind][0],this->mas[ind][1]);
     }
     py::list cumil(){
-        float d;
-        std::vector<float> a((int)this->mas.size(),0);
+        double d;
+        std::vector<double> a((int)this->mas.size(),0);
         for (int i=0;i<(int)this->mas.size()-1;i++){
             d = distance(this->mas[i][0],this->mas[i][1],this->mas[i+1][0],this->mas[i+1][1]);
             a[i+1]=d+a[i];
@@ -83,12 +83,12 @@ class query{
 
 
 py::tuple geo_split(std::string & f){
-    float lt,ln;
+    double lt,ln;
     split(f,&lt,&ln);
     return py::make_tuple(lt,ln);
 }
-float dmstodd(char *a){
-    float t[3],sign=1.0;
+double dmstodd(char *a){
+    double t[3],sign=1.0;
     int j=0,k=0,i=0;
     char temp[8];
     if (a[0]=='-'){sign=-1.0;i=1;}
@@ -104,7 +104,7 @@ float dmstodd(char *a){
     t[k]=atof(temp);
     return (t[0]+t[1]/60+t[2]/3600)*sign;
 }
-void split(std::string & a,float *x,float * y){
+void split(std::string & a,double *x,double * y){
     bool ew=false;
     int dms=0;
     *x=1.0;
@@ -158,7 +158,7 @@ PYBIND11_MODULE(geo, m) {
 }
 
 // int main(){
-//     float a1,b1,a2,b2;
+//     double a1,b1,a2,b2;
 //     std::string f1 ="12.345, -45.455";
 //     std::string f2 ="N12.345E15.455";
 //     split(f1,&a1,&b1);
